Add update-rcc to topgrade

+++ config/dotconfig/fish/functions/update-rcc.fish
@@ -1,7 +1,11 @@
-# Defined in /var/folders/xz/6y6sty192sbg2m3bb623z6hw0000gn/T//fish.fziPu0/update-rcc.fish @ line 2
+# Defined in /tmp/fish.B3QMZI/update-rcc.fish @ line 2
 function update-rcc
+	if test -f $HOME/Code/robocorp/rcc
 		cd /Users/kerkko/Code/robocorp/rcc
 		git pull
 		rake build
 		cp /Users/kerkko/Code/robocorp/rcc/build/macos64/rcc /Users/kerkko/.local/bin/rcc
+	else
+		echo "No RCC repo cloned, skipping update-rcc"
+	end
 end
